+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix: OutOfMemory exception type #1015
+- fix: macOS version for Mac Catalyst #1011
 
 ## 7.0.0-alpha.3